Started restructuring
Everything worked fine, except that touches moving from one control to another one couldn't be properly handled, as every view element was responsible for its own touches. To change that, the GameScene itself should be responsible for any touches. Also, it is planned to remove some classes like CornerButton to try and simplify the structure of the progam as far as possible.
Right now, the game is not working properly, but it's on its way to work better than before. Hopefully.